Overhaul sysdig long option parsing
Long-only options were parsed incorrectly whenever there was
a short option passed after a long-only one (the long-only
handler code was called twice, which was a big issue if it took
an argument, since it would now be passed the short option's
argument).

Fix this by modelling the option parsing code after csysdig
(long options are explicitly handled only when getopt_long
returns 0). This required a fix (with some minor code duplication)
for handling `--list-chisels` as the previous code handled
long-only options in the main `getopt_long()` switch in a pretty
convoluted way.